+++ README.md
@@ -16,7 +16,20 @@
 ## 2/22
 - fs 모듈 이용해 json파일 불러오기 (streaming?)
 - json파일 불러온 것을 html에 넣어보기 
-    -서버를 우선적으로 만들어야할거 같다.
+    -서버를 우선적으로 만들어야할거 같아...서 일단 접음.
 
 ## 2/23
-- createElement, appendChild 제대로 사용하기.
+-  DOM 조작 (1)
+    - querySelector()
+    - querySelectorAll()
+    - createElement
+    - removeChild 
+    - appendChild 
+    - insertBefore
+    - replaceChild
+    - closest
+
+- DOM 조작 (2)
+    - innerText()
+    - innerHTML()
+    - insertAdjacentHTML()
