+++ json_heros_tutorial/memo_JSON_tutorial.txt
@@ -1,0 +1,1 @@
+https://developer.mozilla.org/en-US/docs/Learn/JavaScript/Objects/JSON
