+++ kakao_clone_02/components/dummy.js
@@ -1,0 +1,15 @@
+var dummy = {
+    "Genre":"Action",
+    "Comics":[
+        "Million tonne punch",
+        "Damage resistance",
+        "Superhumann reflexes"
+    ]
+};
+var days = [{
+    "Mon":"Action",
+    "Tue":"Drama",
+    "Wed":"Webtoon",
+    "Thu":"Romance",
+    "Fri":"Fantasy"
+  }];
