+++ kakao_clone_02/components/dummy.json
