+++ kakao_clone_02/index.html
@@ -2,35 +2,23 @@
 <head>
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1">
+  <!-- <link rel="stylesheet" href="./style/test.css"> -->
   <link rel="stylesheet" href="./style/style.css">
-  <script src="./components/main.js"></script>
 </head>
 <body>
-<header class="header">
-  <nav class="nav"></nav>
-  <!-- <div class="header-nav">
-      <h3 class="header-h3" >Kakaopage</h3>
-      <div class="header-menu">
-          <input class="header-input" type="text">
-          <button class="header-stage">Stage</button>
-          <button class="header-mini-menu">
-            <span class = "header-charge">캐시충전</span>
-            
-            <span class = "header-login">로그인</span>
-          </button>
-      </div>
-    </div> -->
+  <header>
+      <div>header(html)</div>
+      <div>page(html)</div>
+    <!-- <nav> </nav> -->
   </header>
-
-  <main class="main">
-    <div class="category-content">
-
-
-    </div>
+  <main>
+      <div class="category-content">
+        <h3>category-content(html)</h3>
+      </div>
   </main>
-
-  <footer class="footer">
-
+  <footer>
   </footer>
+  <script src="./components/main.js"></script>
+  <!-- <script src="./components/test.js"></script> -->
 </body>
 </html>
