+++ kakao_clone_02/style/style.css
@@ -3,31 +3,19 @@
 @import "footer.css"; */
 
 :root {
-  --yellow: #ffec42;
-  --white: #ffffff;
-  --font-dark: rgb(0, 0, 0);
-  --hover-yellow: #fbffd5;
-}
-* {
-  box-sizing: border-box;
-  font-family: NotoSans;
-  line-height: 1.4;
-}
-div {
-  background-color: black;
-}
-/* Set height of body and the document to 100% */
-body, html {
-  background-color: #f1f1f1;
-  width: 100%;
-  height: 100%;
-  margin: 0;
-  font-family: Arial;
-}
-.full-box {
-  -moz-box-sizing: border-box;
-  -webkit-box-sizing: border-box;
-  box-sizing: border-box;
-  width: 100%;
-  height: 100%;
-}
+    --yellow: #ffec42;
+    --white: #ffffff;
+    --font-dark: rgb(0, 0, 0);
+    --hover-yellow: #fbffd5;
+  }
+  * {
+    box-sizing: border-box;
+    font-family: NotoSans;
+    line-height: 1.4;
+  }
+  body, html {
+    background-color: #f1f1f1;
+    width: 100%;
+    height: 100%;
+    margin: 0;
+  }
