+++ kakao_clone_02/style/test.css
@@ -1,0 +1,43 @@
+/* @import "main.css";
+@import "header.css";
+@import "footer.css"; */
+
+:root {
+  --yellow: #ffec42;
+  --white: #ffffff;
+  --font-dark: rgb(0, 0, 0);
+  --hover-yellow: #fbffd5;
+}
+* {
+  box-sizing: border-box;
+  font-family: NotoSans;
+  line-height: 1.4;
+}
+/* Set height of body and the document to 100% */
+body, html {
+  background-color: #f1f1f1;
+  width: 100%;
+  height: 100%;
+  margin: 0;
+}
+header {
+  display: flex;
+  align-items: center;
+  flex-direction: row;
+}
+main {
+  display: flex;
+  align-items: center;
+  flex-direction: column;
+}
+div {
+  background-color: black;
+  color: var(--white);
+  width: 300px;
+  height: 50px;
+  margin: 10px;
+}
+p {
+  background-color: #ffec42;
+  margin: 10px;
+}
