+++ memo_querySelector.txt
@@ -1,6 +1,4 @@
-https://developer.mozilla.org/en-US/docs/Web/API/Document/querySelector
-
-# Document.querySelector()
+# Document.querySelector() [source](https://developer.mozilla.org/en-US/docs/Web/API/Document/querySelector)
 
 The Document method querySelector() returns the first Element within the document that matches the specified selector, or group of selectors. If no matches are found, null is returned.
 
@@ -19,7 +17,6 @@ var el = document.querySelector(".myclass");
 
 =======================================
 If you need a list of all elements matching the specified selectors, you should use querySelectorAll() instead.
-https://developer.mozilla.org/en-US/docs/Web/API/Document/querySelectorAll
 
-# Document.querySelectorAll()
+# Document.querySelectorAll() [source](https://developer.mozilla.org/en-US/docs/Web/API/Document/querySelectorAll)
